Make directories and move files to conform to other EUPS packages
