+++ SConstruct
@@ -1,0 +1,3 @@
+# -*- python -*-
+from lsst.sconsUtils import scripts
+scripts.BasicSConstruct("drp_fstar", versionModuleName="python/pfs/drp/fstar/version.py", disableCc=True)
