+++ bin.src/makeRBFInterpolationModel.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+from pfs.drp.fstar.makeRBFInterpolationModel import main
+main()
