+++ setup.cfg
@@ -1,0 +1,13 @@
+[flake8]
+max-line-length = 110
+ignore = E133, E226, E228, E266, N802, N803, N806, N812, N813, N815, N816, W504
+# TODO: remove E266 when Task documentation is converted to rst
+exclude =
+    bin/*
+    sphinx/conf.py
+    __init__.py
+    tests/.tests/*
+
+[tool:pytest]
+addopts = --flake8
+flake8-ignore = E133 E226 E228 N802 N803 N806 N812 N813 N815 N816 W504
