+++ ups/drp_fstar.cfg
@@ -1,0 +1,10 @@
+# -*- python -*-
+
+import lsst.sconsUtils
+
+dependencies = {}
+
+config = lsst.sconsUtils.Configuration(
+    __file__,
+    hasDoxygenInclude=False,
+)
