+++ ups/drp_fstar.table
@@ -1,0 +1,5 @@
+setupRequired(sconsUtils)
+setupRequired(utils)
+
+envPrepend(PATH, ${PRODUCT_DIR}/bin)
+envPrepend(PYTHONPATH, ${PRODUCT_DIR}/python)
